+++ mcu_project_msp430fr2433/i2c.h
@@ -27,13 +27,13 @@ enum eI2C_MODE
     I2C_TIMEOUT_MODE
 };
 
-enum eI2C_MODE i2c_write(uint8_t dev_addr, uint8_t *data, uint8_t count);
-enum eI2C_MODE i2c_write_reg1(uint8_t dev_addr, uint8_t reg, uint8_t *data, uint8_t count);
-enum eI2C_MODE i2c_write_reg2(uint8_t dev_addr, uint16_t reg, uint8_t *data, uint8_t count);
+enum eI2C_MODE i2c_write(uint8_t dev_addr, uint8_t *data, uint16_t count);
+enum eI2C_MODE i2c_write_reg1(uint8_t dev_addr, uint8_t reg, uint8_t *data, uint16_t count);
+enum eI2C_MODE i2c_write_reg2(uint8_t dev_addr, uint16_t reg, uint8_t *data, uint16_t count);
 
-enum eI2C_MODE i2c_read(uint8_t dev_addr, uint8_t* dest, uint8_t count);
-enum eI2C_MODE i2c_read_reg1(uint8_t dev_addr, uint8_t reg, uint8_t* dest, uint8_t count);
-enum eI2C_MODE i2c_read_reg2(uint8_t dev_addr, uint16_t reg, uint8_t* dest, uint8_t count);
+enum eI2C_MODE i2c_read(uint8_t dev_addr, uint8_t* dest, uint16_t count);
+enum eI2C_MODE i2c_read_reg1(uint8_t dev_addr, uint8_t reg, uint8_t* dest, uint16_t count);
+enum eI2C_MODE i2c_read_reg2(uint8_t dev_addr, uint16_t reg, uint8_t* dest, uint16_t count);
 
 bool i2c_got_nack(void);
 void i2c_init(void);
